+++ openvpn/src/main/cpp/openvpn3/openvpn/tun/builder/capture.hpp
@@ -4,7 +4,7 @@
 //               packet encryption, packet authentication, and
 //               packet compression.
 //
-//    Copyright (C) 2012-2017 OpenVPN Inc.
+//    Copyright (C) 2012-2020 OpenVPN Inc.
 //
 //    This program is free software: you can redistribute it and/or modify
 //    it under the terms of the GNU Affero General Public License Version 3
@@ -787,7 +787,7 @@ namespace openvpn {
     {
       if (index == -1)
 	return true;
-      return index >= 0 && index <= tunnel_addresses.size();
+      return index >= 0 && static_cast<unsigned int>(index) <= tunnel_addresses.size();
     }
 
     void validate_tunnel_address_indices(const std::string& title) const
